+++ GrowthBook/src/commonMain/kotlin/com/sdk/growthbook/GrowthBookSDK.kt
@@ -268,7 +268,7 @@ class GrowthBookSDK internal constructor(
     /**
      * Get Cached Features
      */
-    fun getFeatures(): GBFeatures {
+    override fun getFeatures(): GBFeatures {
         return gbContext.features
     }